sample: psa_level_1: PSA level 1 test app
This commit adds a test app showing how the initial attestation, secure
storage and crypto services from TF-M can be used on the non-secure (NS)
environment, making calls to the secure service implementations via
veneer function calls.

It is based on TF-M version 1.0-rc2, and supersedes #18185.

This PR makes use of a common build infrastructure to pull in the TF-M
repo and dependencies, and merge and sign the secure TF-M and NS Zephyr
binaries during the standard Zephyr build process.

Signed-off-by: Kevin Townsend <kevin@ktownsend.com>